requirements.txt
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
fastapi==0.68.1
flake8==3.9.2
h11==0.12.0
importlib-metadata==4.7.0
mccabe==0.6.1
Pillow==8.4.0
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
Pygments==2.10.0
pytz==2021.3
sqlparse==0.4.2
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
zipp==3.5.0
